requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contextlib2==21.6.0
cryptography==46.0.3
Django>=5.2
django-mptt==0.17.0
django-ninja
django-ninja-extra==0.30.2
django-ninja-jwt==5.4.0
idna==3.11
injector==0.22.0
psycopg==3.2.12
psycopg-binary==3.2.12
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==12.0
